wire: fleshing awaiting_channel_range and awaiting_short_ids states
